requirements.txt
django
djangorestframework
django-rest-knox
django-cors-headers
